refactor: cleanup quick wins (dead code, typo, type location)
- Delete unused EpisodeStatistics/LibraryStatistics components (zero refs)
- Rename extractPhrasesUitls.ts -> extractPhrasesUtils.ts (fix typo) + imports
- Rename useSpacedRepetitionGame.tsx -> .ts (contains no JSX)
- Move PhraseItem interface from AnkiExporter.tsx to types/phrase.ts
