+++ src/types/phrase.ts
@@ -1,0 +1,5 @@
+export interface PhraseItem {
+  phrase: string;
+  translation: string;
+  frequency?: number;
+}
