+++ src/utils/extractPhrasesUtils.ts
